Aula-RAG/requirements.txt
openai>=1.50.0
tiktoken
numpy
pandas
scikit-learn
gradio
python-dotenv
